Fix: rename extensionContext file to lower case
