+++ src/lib/extensionContext.ts
