+++ src/lib/firmHandler.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import * as utils from "../utilities/utils";
-import ExtensionContext from "./ExtensionContext";
+import ExtensionContext from "./extensionContext";
 import ExtensionLogger from "./outputChannels/extensionLogger";
 import StatusBarItem from "./statusBar/statusBarItem";
 const { firmCredentials } = require("silverfin-cli/lib/api/firmCredentials");
